Fixed dashboard.
Fixed the dashboard so that when there are no data in the database, the webapp will still run.

This change checks if feedbackData exists and is an array before attempting to map over it. If it doesn't meet these conditions, it sets allServices to ["All"] as a default value. This will prevent the error when feedbackData is empty or not yet loaded. This also goes for the .map functions and the .length functions.